requirements.txt
annotated-types==0.6.0
asgiref==3.8.1
beautifulsoup4==4.9.3
bert-extractive-summarizer==0.10.1
blis==0.7.11
bs4==0.0.1
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==2.0.12
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
confection==0.1.4
cssselect==1.2.0
cymem==2.0.8
Django==5.0.3
django-social-share==2.3.0
dnspython==1.16.0
feedfinder2==0.0.4
feedparser==6.0.11
filelock==3.13.3
fsspec==2024.3.1
gnews==0.3.6
gunicorn==21.2.0
huggingface-hub==0.22.2
idna==3.6
jieba3k==0.35.1
Jinja2==3.1.3
joblib==1.3.2
langcodes==3.3.0
lxml==5.1.1
MarkupSafe==2.1.5
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pillow==10.2.0
preshed==3.0.9
protobuf==5.26.1
pydantic==2.6.4
pydantic_core==2.16.3
pymongo==3.12.3
python-dateutil==2.9.0.post0
python-dotenv==0.19.2
pytz==2024.1
PyYAML==6.0.1
rapidfuzz==3.7.0
regex==2023.12.25
requests
requests-file
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
sentencepiece==0.2.0
setuptools==69.2.0
sgmllib3k==1.0.0
six==1.16.0
smart-open==6.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.4.4
srsly==2.4.8
sympy==1.12
thinc==8.2.3
threadpoolctl==3.4.0
tinysegmenter==0.3
tldextract==5.1.2
tokenizers==0.15.2
torch==2.2.2
tqdm==4.66.2
transformers==4.39.2
typer==0.9.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==1.26.18
wasabi==1.1.2
weasel==0.3.4
